+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Scroll to top of the list when clicking on toolbar
 - Thin item dividers for all lists
 - Confirmation dialogs for destructive actions (logout and item deletion)
+- Confirmation dialog for discard item and item authorization changes action
 
 ## Changed
 - Allow to copy item details of read-only items (do not disable them anymore) - the delete button is hidden instead of disabled to avoid user confusion (like the save accept toolbar icon)